#11 step2: supporting endpoint specification though the uri e.g. s3://[[key:secret@]endpoint]/bucket[/resource]
