+++ src/test/java/software/amazon/nio/spi/s3/S3FileSystemTest.java
@@ -109,59 +109,4 @@ public class S3FileSystemTest {
     public void testGetOpenChannelsIsNotModifiable() {
         s3FileSystem.getOpenChannels().add(null);
     }
-
-    @Test
-    public void clientsWithProvidedEndpoint() throws Exception {
-        when(mockClient.getBucketLocation(any(Consumer.class)))
-                .thenReturn(GetBucketLocationResponse.builder().locationConstraint("us-west-2").build());
-        S3FileSystemProvider.getClientStore().locationClient(mockClient);
-
-        final String URI1 = "s3://endpoint1.io/bucket/resource";
-        final String URI2 = "s3://endpoint2.io:8080/bucket/resource";
-
-        S3FileSystem fs = new S3FileSystem(URI.create(URI1), provider);
-        S3Client client = fs.client();
-
-        Field f = client.getClass().getDeclaredField("clientConfiguration");
-        f.setAccessible(true);
-        SdkClientConfiguration sdkConf = (SdkClientConfiguration)f.get(client);
-
-        assertTrue(sdkConf.option(ENDPOINT_OVERRIDDEN));
-        assertEquals(URI.create("https://endpoint1.io"), sdkConf.option(ENDPOINT));
-
-        sdkConf = (SdkClientConfiguration)f.get(new S3FileSystem(URI.create(URI2), provider).client());
-        assertTrue(sdkConf.option(ENDPOINT_OVERRIDDEN));
-        assertEquals(URI.create("https://endpoint2.io:8080"), sdkConf.option(ENDPOINT));
-    }
-
-    @Test
-    public void clientsWithProvidedEndpointAndCredentials() throws Exception {
-        when(mockClient.getBucketLocation(any(Consumer.class)))
-                .thenReturn(GetBucketLocationResponse.builder().locationConstraint("us-west-2").build());
-        S3FileSystemProvider.getClientStore().locationClient(mockClient);
-
-        final String URI1 = "s3://key1:secret1@endpoint1.io/bucket/resource";
-        final String URI2 = "s3://key2:secret2@endpoint2.io:8080/bucket/resource";
-
-        S3FileSystem fs = new S3FileSystem(URI.create(URI1), provider);
-        S3Client client = fs.client();
-
-        Field f = client.getClass().getDeclaredField("clientConfiguration");
-        f.setAccessible(true);
-        SdkClientConfiguration sdkConf = (SdkClientConfiguration)f.get(client);
-
-        assertEquals(URI.create("https://endpoint1.io"), sdkConf.option(ENDPOINT));
-        assertFalse(sdkConf.option(CREDENTIALS_PROVIDER) instanceof DefaultCredentialsProvider);
-        AwsCredentialsProvider credentials = (AwsCredentialsProvider)sdkConf.option(CREDENTIALS_PROVIDER);
-        assertEquals("key1", credentials.resolveCredentials().accessKeyId());
-        assertEquals("secret1", credentials.resolveCredentials().secretAccessKey());
-
-
-        sdkConf = (SdkClientConfiguration)f.get(new S3FileSystem(URI.create(URI2), provider).client());
-        assertEquals(URI.create("https://endpoint2.io:8080"), sdkConf.option(ENDPOINT));
-        credentials = (AwsCredentialsProvider)sdkConf.option(CREDENTIALS_PROVIDER);
-        assertEquals("key2", credentials.resolveCredentials().accessKeyId());
-        assertEquals("secret2", credentials.resolveCredentials().secretAccessKey());
-    }
-
 }
